requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
py-mini-racer==0.2.0
slackclient==2.7.1
sq-native==0.6.1.0.1
sqreen==1.20.0
Werkzeug==1.0.1
yarl==1.4.2
